refactor: move rig state snapshot types to homeboy-lifecycle-contract
Clears the last TraceResults blocker. RigStateSnapshot + ComponentSnapshot are
pure serde state-snapshot types (ComponentSnapshot: 0 impls; RigStateSnapshot: 1
pure impl set_effective_component_path that takes a git-lookup closure and
touches no core APIs). They move to the lifecycle/evidence-contract leaf crate
alongside the artifact/timeline evidence types.

MaterializedRigState (which carries a BTreeMap<String, ComponentSnapshot>) and
the other rig state types stay in core and re-import ComponentSnapshot. Core's
rig module re-exports both moved types so all consumers (rig, trace, bench, cli)
stay stable.

With this, both TraceResults blockers (TracePreviewMetadata #8841 and now
RigStateSnapshot) are cleared - the top-level trace result cluster can move next.
